Move static action blocks to another module
